Make product's put route require all fields
Initially it was a 'patch' route, since the task requires the 'put'
route, I changed the 'patch' to 'put' but the optional fields stayed.
Now I think that I have to make it a 'real' 'put' route, so the
task will be fullfilled. Although I think that it should be a 'patch'
route and allow optional fields.
The tests are changed accordingly too.